+++ txros/src/txros/__init__.py
@@ -5,6 +5,6 @@ from .rosxmlrpc import Error, Proxy
 from .serviceclient import ServiceError, ServiceClient
 from .service import Service
 from .subscriber import Subscriber
-from .tf import TransformBroadcaster, TooPastError, TransformListener, Transform
+from .txros_tf import TransformBroadcaster, TooPastError, TransformListener, Transform
 from .variable import Event, Variable
 from .util import DeferredCancelDeferred
